% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{extract_paged_data}
\alias{extract_paged_data}
\title{Extract Data from Paged Response}
\usage{
extract_paged_data(resp, key)
}
\arguments{
\item{resp}{A list containing the API response.}

\item{key}{The key containing the data (e.g., "occupation", "rows").}
}
\value{
A tibble of the extracted data.
}
\description{
Extracts the data portion from a paged API response.
}
\keyword{internal}
