% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{is_transient_error}
\alias{is_transient_error}
\title{Check if HTTP Response is a Transient Error}
\usage{
is_transient_error(resp)
}
\arguments{
\item{resp}{An httr2 response object.}
}
\value{
Logical indicating if the error is transient.
}
\description{
Determines if an HTTP response represents a transient error that should be retried.
}
\keyword{internal}
