% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupations.R
\name{onet_apprenticeship}
\alias{onet_apprenticeship}
\title{Get O*NET Apprenticeship Opportunities (details)}
\usage{
onet_apprenticeship(code, start = 1, end = 20)
}
\arguments{
\item{code}{An O*NET-SOC occupation code (e.g., "15-1252.00").}

\item{start}{Integer specifying the first result to return (default 1).}

\item{end}{Integer specifying the last result to return (default 20).}
}
\value{
A tibble with one row per apprenticeship example title.
}
\description{
Get O*NET Apprenticeship Opportunities (details)
}
