% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupations.R
\name{onet_detailed_work_activities}
\alias{onet_detailed_work_activities}
\title{Get O*NET Detailed Work Activities (details)}
\usage{
onet_detailed_work_activities(code, start = 1, end = 20)
}
\description{
Note: this endpoint returns items under \code{activity} with fields like id/title/related.
}
