% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupations.R
\name{onet_interests}
\alias{onet_interests}
\title{Get O*NET Occupation Interests (details)}
\usage{
onet_interests(code, start = 1, end = 20)
}
\arguments{
\item{code}{An O*NET-SOC occupation code.}

\item{start}{Integer specifying the first result to return.}

\item{end}{Integer specifying the last result to return.}
}
\value{
A tibble of interest elements.
}
\description{
Get O*NET Occupation Interests (details)
}
\examples{
\dontshow{if (nzchar(Sys.getenv("ONET_API_KEY"))) withAutoprint(\{ # examplesIf}
onet_interests("15-1252.00", end = 5)
\dontshow{\}) # examplesIf}
}
