% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupations.R
\name{onet_knowledge}
\alias{onet_knowledge}
\title{Get O*NET Occupation Knowledge (details)}
\usage{
onet_knowledge(code, start = 1, end = 20)
}
\arguments{
\item{code}{An O*NET-SOC occupation code.}

\item{start}{Integer specifying the first result to return.}

\item{end}{Integer specifying the last result to return.}
}
\value{
A tibble of knowledge elements.
}
\description{
Get O*NET Occupation Knowledge (details)
}
