% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupations.R
\name{onet_occupations}
\alias{onet_occupations}
\title{List All O*NET Occupations}
\usage{
onet_occupations(start = 1, end = 1000)
}
\arguments{
\item{start}{Integer specifying the first result to return (default 1).}

\item{end}{Integer specifying the last result to return (default 1000).}
}
\value{
A tibble with columns:
\describe{
\item{code}{O*NET-SOC occupation code}
\item{title}{Occupation title}
}
}
\description{
Retrieves a list of all occupations in the O*NET database.
}
\examples{
\dontrun{
occupations <- onet_occupations()
head(occupations)
}
}
