% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupations.R
\name{onet_related_occupations}
\alias{onet_related_occupations}
\title{Get O*NET Related Occupations (details)}
\usage{
onet_related_occupations(code, start = 1, end = 20)
}
\arguments{
\item{code}{An O*NET-SOC occupation code.}

\item{start}{Integer specifying the first result to return.}

\item{end}{Integer specifying the last result to return.}
}
\value{
A tibble of related occupations.
}
\description{
Get O*NET Related Occupations (details)
}
