% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupations.R
\name{onet_skills}
\alias{onet_skills}
\title{Get O*NET Occupation Skills (details)}
\usage{
onet_skills(code, start = 1, end = 20)
}
\arguments{
\item{code}{An O*NET-SOC occupation code.}

\item{start}{Integer specifying the first result to return.}

\item{end}{Integer specifying the last result to return.}
}
\value{
A tibble of skills.
}
\description{
Get O*NET Occupation Skills (details)
}
\examples{
\dontshow{if (interactive() && nzchar(Sys.getenv("ONET_API_KEY"))) withAutoprint(\{ # examplesIf}
onet_skills("15-1252.00", end = 5)
\dontshow{\}) # examplesIf}
}
