% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{onet_tables}
\alias{onet_tables}
\title{List O\*NET Database Tables}
\usage{
onet_tables()
}
\value{
A tibble with columns:
\describe{
\item{id}{Table identifier}
\item{title}{Table title/description}
}
}
\description{
Retrieves a list of available database tables in O\*NET.
}
\examples{
\dontshow{if (nzchar(Sys.getenv("ONET_API_KEY"))) withAutoprint(\{ # examplesIf}
tables <- onet_tables()
head(tables)
\dontshow{\}) # examplesIf}
}
