% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupations.R
\name{onet_work_activities}
\alias{onet_work_activities}
\title{Get O*NET Occupation Work Activities (details)}
\usage{
onet_work_activities(code, start = 1, end = 20)
}
\description{
Get O*NET Occupation Work Activities (details)
}
