% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupations.R
\name{onet_work_context}
\alias{onet_work_context}
\title{Get O*NET Occupation Work Context (details)}
\usage{
onet_work_context(code, start = 1, end = 20)
}
\description{
Get O*NET Occupation Work Context (details)
}
