% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupations.R
\name{onet_work_styles}
\alias{onet_work_styles}
\title{Get O*NET Occupation Work Styles (details)}
\usage{
onet_work_styles(code, start = 1, end = 20)
}
\arguments{
\item{code}{An O*NET-SOC occupation code.}

\item{start}{Integer specifying the first result to return.}

\item{end}{Integer specifying the last result to return.}
}
\value{
A tibble of work style elements.
}
\description{
Get O*NET Occupation Work Styles (details)
}
