% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{to_snake_case}
\alias{to_snake_case}
\title{Convert to Snake Case}
\usage{
to_snake_case(x)
}
\arguments{
\item{x}{Character vector to convert.}
}
\value{
Character vector in snake_case.
}
\description{
Convert to Snake Case
}
\keyword{internal}
