% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{as_onet_tibble}
\alias{as_onet_tibble}
\title{Convert API Response to Tibble}
\usage{
as_onet_tibble(x)
}
\arguments{
\item{x}{A list or data frame to convert.}
}
\value{
A tibble.
}
\description{
Converts a list from the API to a tibble with snake_case column names.
}
\keyword{internal}
