% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onet2r-package.R
\docType{package}
\name{onet2r-package}
\alias{onet2r}
\alias{onet2r-package}
\title{onet2r: Access 'O*NET' Web Services API}
\description{
Access the 'O*NET' Web Services Application Programming Interface (API) version 2.0. Search occupations, retrieve occupation details such as skills, knowledge, abilities, tasks, and work activities, access database tables, and perform taxonomy crosswalks. Results are returned as tibbles for easy integration with tidyverse workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/farach/onet2r}
  \item Report bugs at \url{https://github.com/farach/onet2r/issues}
}

}
\author{
\strong{Maintainer}: Alex Farach \email{alexfarach@microsoft.com} [copyright holder]

}
\keyword{internal}
