% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onet2r-package.R
\docType{package}
\name{onet2r-package}
\alias{onet2r}
\alias{onet2r-package}
\title{onet2r: Work with 'O*NET' Web Services, Archives, and Weights}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Access the 'O*NET' Web Services Application Programming Interface (API) version 2.0. Search occupations, retrieve occupation details such as skills, knowledge, abilities, tasks, and work activities, access database tables, perform taxonomy crosswalks, join wage and employment context, assemble longitudinal archive panels, and validate user-supplied occupational measures. Use archived releases, source dates, and taxonomy bridges to make longitudinal comparisons easier to audit. Results are returned as tibbles for easy integration with tidyverse workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://farach.github.io/onet2r/}
  \item \url{https://github.com/farach/onet2r}
  \item Report bugs at \url{https://github.com/farach/onet2r/issues}
}

}
\author{
\strong{Maintainer}: Alex Farach \email{alexfarach@microsoft.com} [copyright holder]

}
\keyword{internal}
