% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onet2r-package.R
\docType{package}
\name{onet2r-package}
\alias{onet2r}
\alias{onet2r-package}
\title{onet2r: Access O*NET Web Services API}
\description{
An R client for the O*NET Web Services API 2.0. Provides functions to search occupations, retrieve occupation details (skills, knowledge, abilities), access database tables, and perform taxonomy crosswalks. All functions return tibbles for easy integration with tidyverse workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alexfarach/onet2r}
  \item Report bugs at \url{https://github.com/alexfarach/onet2r/issues}
}

}
\author{
\strong{Maintainer}: Alex Farach \email{alex@example.com}

}
\keyword{internal}
