% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupations.R
\name{onet_detailed_work_activities}
\alias{onet_detailed_work_activities}
\title{Get O*NET Detailed Work Activities (details)}
\usage{
onet_detailed_work_activities(code, start = 1, end = 20)
}
\arguments{
\item{code}{An O*NET-SOC occupation code.}

\item{start}{Integer specifying the first result to return.}

\item{end}{Integer specifying the last result to return.}
}
\value{
A tibble of detailed work activities.
}
\description{
Note: this endpoint returns items under \code{activity} with fields like id/title/related.
}
