% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupations.R
\name{onet_in_demand_skills}
\alias{onet_in_demand_skills}
\title{Get O*NET In-Demand Skills (details)}
\usage{
onet_in_demand_skills(code, start = 1, end = 20)
}
\description{
Get O*NET In-Demand Skills (details)
}
