% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{onet_measure_coverage}
\alias{onet_measure_coverage}
\title{Return Measure Coverage}
\usage{
onet_measure_coverage(measure)
}
\arguments{
\item{measure}{An object from \code{\link[=onet_measure]{onet_measure()}}.}
}
\value{
A tibble with key and optional employment coverage.
}
\description{
Return Measure Coverage
}
