% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupations.R
\name{onet_occupation}
\alias{onet_occupation}
\title{Get O*NET Occupation Overview}
\usage{
onet_occupation(code)
}
\arguments{
\item{code}{An O*NET-SOC occupation code (e.g., "15-1252.00").}
}
\value{
A list containing the occupation overview data.
}
\description{
Retrieves overview information for a specific occupation.
}
