% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupations.R
\name{onet_professional_associations}
\alias{onet_professional_associations}
\title{Get O*NET Professional Associations (details)}
\usage{
onet_professional_associations(code, start = 1, end = 20)
}
\arguments{
\item{code}{An O*NET-SOC occupation code.}

\item{start}{Integer specifying the first result to return.}

\item{end}{Integer specifying the last result to return.}
}
\value{
A tibble of professional associations.
}
\description{
Get O*NET Professional Associations (details)
}
