% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupations.R
\name{onet_related_occupations}
\alias{onet_related_occupations}
\title{Get O*NET Related Occupations (details)}
\usage{
onet_related_occupations(code, start = 1, end = 20)
}
\description{
Get O*NET Related Occupations (details)
}
