% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{onet_tables}
\alias{onet_tables}
\title{List O*NET Database Tables}
\usage{
onet_tables()
}
\value{
A tibble with columns:
\describe{
\item{id}{Table identifier}
\item{title}{Table title/description}
}
}
\description{
Retrieves a list of available database tables in O*NET.
}
\examples{
\dontrun{
tables <- onet_tables()
head(tables)
}
}
