% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupations.R
\name{onet_tasks}
\alias{onet_tasks}
\title{Get O*NET Occupation Tasks (details)}
\usage{
onet_tasks(code, start = 1, end = 20)
}
\arguments{
\item{code}{An O*NET-SOC occupation code.}

\item{start}{Integer specifying the first result to return.}

\item{end}{Integer specifying the last result to return.}
}
\value{
A tibble of tasks.
}
\description{
Get O*NET Occupation Tasks (details)
}
