% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupations.R
\name{onet_work_styles}
\alias{onet_work_styles}
\title{Get O*NET Occupation Work Styles (details)}
\usage{
onet_work_styles(code, start = 1, end = 20)
}
\description{
Get O*NET Occupation Work Styles (details)
}
